requirements.txt
PyQt5~=5.15.7
PyYAML~=6.0
pandas~=1.3.5
numpy~=1.21.6
opencv-python~=4.6.0.66
Pillow~=9.3.0
six~=1.16.0
scikit-image~=0.19.3
scipy~=1.7.3
tornado~=6.2
setuptools~=57.0.0
